+++ introspection-engine/connectors/mongodb-introspection-connector/tests/errors/mod.rs
@@ -1,4 +1,4 @@
-use crate::common::*;
+use crate::test_api::*;
 use mongodb_introspection_connector::MongoDbIntrospectionConnector;
 
 #[test]
