+++ introspection-engine/connectors/mongodb-introspection-connector/tests/index/mod.rs
@@ -1,4 +1,4 @@
-use crate::common::*;
+use crate::test_api::*;
 use bson::Bson;
 use mongodb::{options::IndexOptions, IndexModel};
