+++ introspection-engine/connectors/mongodb-introspection-connector/tests/lib.rs
@@ -1,4 +1,4 @@
-mod common;
+mod test_api;
 
 mod basic;
 mod dirty_data;
