+++ introspection-engine/connectors/mongodb-introspection-connector/tests/model_renames/mod.rs
@@ -1,4 +1,4 @@
-use crate::common::*;
+use crate::test_api::*;
 
 #[test]
 fn a_model_with_reserved_name() {
