+++ introspection-engine/connectors/mongodb-introspection-connector/tests/types/mod.rs
@@ -1,4 +1,4 @@
-use crate::common::*;
+use crate::test_api::*;
 use bson::{oid::ObjectId, Array, Binary, Bson, DateTime, Decimal128, Timestamp};
 
 #[test]
